feat: Migrate the module to typescript
