+++ Scripts/build-webdriveragent.mjs
@@ -1,8 +1,13 @@
-const path = require('path');
-const { asyncify } = require('asyncbox');
-const { logger, fs } = require('@appium/support');
-const { exec } = require('teen_process');
-const xcode = require('appium-xcode');
+import path from 'node:path';
+import { fileURLToPath } from 'node:url';
+import { asyncify } from 'asyncbox';
+import { logger, fs } from '@appium/support';
+import { exec } from 'teen_process';
+import xcode from 'appium-xcode';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+const isMainModule = process.argv[1] && path.resolve(process.argv[1]) === __filename;
 
 const LOG = new logger.getLogger('WDABuild');
 const ROOT_DIR = path.resolve(__dirname, '..');
@@ -72,8 +77,9 @@ async function buildWebDriverAgent (xcodeVersion) {
   LOG.info(`Zip bundled at "${appBundleZipPath}"`);
 }
 
-if (require.main === module) {
+if (isMainModule) {
   asyncify(buildWebDriverAgent);
 }
 
-module.exports = buildWebDriverAgent;
+export default buildWebDriverAgent;
+
