+++ Scripts/update-wda-version.mjs
@@ -1,6 +1,10 @@
-const {plist, logger} = require('@appium/support');
-const path = require('node:path');
-const semver = require('semver');
+import {plist, logger} from '@appium/support';
+import path from 'node:path';
+import { fileURLToPath } from 'node:url';
+import semver from 'semver';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
 
 const log = logger.getLogger('Versioner');
 
@@ -39,3 +43,4 @@ async function updateWdaVersion() {
 }
 
 (async () => await updateWdaVersion())();
+
