+++ index.ts
@@ -1,4 +1,4 @@
-export { checkForDependencies, bundleWDASim } from './lib/check-dependencies';
+export { bundleWDASim } from './lib/check-dependencies';
 export { NoSessionProxy } from './lib/no-session-proxy';
 export { WebDriverAgent } from './lib/webdriveragent';
 export { WDA_BASE_URL, WDA_RUNNER_BUNDLE_ID, PROJECT_FILE } from './lib/constants';
