--- lib/logger.js
@@ -1,5 +1,0 @@
-import { logger } from '@appium/support';
-
-const log = logger.getLogger('WebDriverAgent');
-
-export default log;
