+++ lib/logger.ts
@@ -1,0 +1,3 @@
+import { logger } from '@appium/support';
+
+export const log = logger.getLogger('WebDriverAgent');
